[VSC-7] Support cache in the extension
Problem: We'd like to speed up the extension by adding a cache for
unchanged files.

Solution: Use the new `parseMotokoTypedLsp` function, store its cache,
and give back this cache to the function on each typed run.

Note: The `notify` function is called twice, which severely degrades
performance. The extraneous call is removed.