[VSC-7] Support cache in the extension
Problem: We'd like to speed up the extension by adding a cache for
unchanged files.

Solution: For the typed run, use the new `parseMotokoTypedLsp` function,
store its cache, and give back this cache to the function on each typed
run. For the untyped run, use the new `parseMotokoWithDeps`. For both
cases, manage a build graph and use it to figure out when to remove
things from the cache.