[VSC-7] Fix errors on initialization
Problem: There are some errors during the initialization of the
extension due `parseMotokoWithDeps` failing to analyze dependencies due
to the virtual file system not being completed yet. Althoguh these
errors do not worsen user interface, it's best to solve them.

Solution: Pass a flag indicating whether to get dependencies or not to
the AST resolver. This is controlled based on whether the virtual
filesystem is being loaded or not (using a new flag). Expose a request
to get the dependency graph and write a test using it.

Also, I now log any errors that might happen from the AST resolver.